Clear the clang-tidy findings the new view headers introduced
Reproduced the Clang-Tidy job locally rather than waiting for it: the generated
per-header units, clang-tidy with this repository's own config, WarningsAsErrors
on. Four findings, all in code added by the previous two commits.

misc-unconventional-assign-operator, twice on array_reference::operator=. It
wants the ordinary shape, a non-const member returning a mutable reference, and
a proxy reference is not that: the assignment writes through the proxy rather
than to it, so the proxy itself need not be modifiable. C++23 gives
std::vector<bool>::reference this exact overload -- const-qualified, returning a
const reference -- added by P2321R2 so a prvalue proxy out of a view's
operator* stays assignable. The check reports the standard's own shape as
unconventional, so both overloads carry a NOLINT saying why.

readability-implicit-bool-conversion in array_view's operator<=>. This one is
real: bool <=> bool promotes both operands, and the comparator now says
static_cast<int>(x) <=> static_cast<int>(y) rather than leaving it to the
promotion. False before true either way.

modernize-use-nodiscard on set_view::erase, which the check reaches only because
a view's modifiers are const members, writing through the pointer rather than to
it. Not marked, for the reason std::set::erase is not: the count is there for
callers who want it and discarded by the ones who don't.

Two findings remain, both performance-move-const-arg in finite_bit_set.hpp.
Those lines are byte-identical to main's, where CI is green, so the version of
clang-tidy the job runs does not report them and the copy to hand does. Left
alone rather than changed on the say-so of a tool the workflow does not use.

Also checked the other gate these headers newly pass through: the self-
sufficiency units compile with -Weverything -Werror -pedantic-errors and this
repository's suppression list, and produce no warnings. All four view tests and
18 of 21 standalone header compiles still pass, unchanged.

+++ include/xstd/bits/ranges/array_view.hpp
@@ -255,14 +255,22 @@ public:
                 return array_find<Bits>::at(m_ref, m_idx);
         }
 
-        constexpr array_reference const& operator=(bool value) const noexcept
+        // const-qualified and returning a const reference, which is what a proxy
+        // reference is: the assignment writes through the proxy rather than to
+        // it, so the proxy itself need not be modifiable. C++23 gives
+        // std::vector<bool>::reference exactly this overload, added by P2321R2
+        // so that a prvalue proxy coming out of a view's operator* is still
+        // assignable. misc-unconventional-assign-operator only knows the
+        // ordinary shape - a non-const member returning a mutable reference -
+        // and reports the standard's own proxy shape as unconventional.
+        constexpr array_reference const& operator=(bool value) const noexcept  // NOLINT(misc-unconventional-assign-operator)
                 requires (not IsConst) and requires(Bits& c) { array_ops<Bits>::assign(c, 0UZ, true); }
         {
                 array_ops<Bits>::assign(m_ref, m_idx, value);
                 return *this;
         }
 
-        constexpr array_reference const& operator=(array_reference const& other) const noexcept
+        constexpr array_reference const& operator=(array_reference const& other) const noexcept  // NOLINT(misc-unconventional-assign-operator)
                 requires (not IsConst) and requires(Bits& c) { array_ops<Bits>::assign(c, 0UZ, true); }
         {
                 return *this = static_cast<bool>(other);
@@ -375,7 +383,7 @@ public:
         {
                 return std::lexicographical_compare_three_way(
                         lhs.begin(), lhs.end(), rhs.begin(), rhs.end(),
-                        [](bool x, bool y) static noexcept { return x <=> y; });
+                        [](bool x, bool y) static noexcept { return static_cast<int>(x) <=> static_cast<int>(y); });
         }
 };
